chore(): updates for pre-V5 deprecations
